backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-socketio==5.10.0
python-dotenv==1.0.0
boto3==1.29.0
fastapi-cors==0.0.6
